examples/python-plugin-minimal/requirements.txt
eclipse-zenoh>=1.0.0
pyyaml>=6.0
